// packages/shared/src/types.d.ts
export declare const OKLCHSchema: z.ZodObject<{
    l: z.ZodNumber;
    c: z.ZodNumber;
    h: z.ZodNumber;
    alpha: z.ZodOptional<z.ZodNumber>;
}, z.core.$strip>;
// ...
export declare const ColorValueSchema: z.ZodObject<{
    name: z.ZodString;
    scale: z.ZodArray<z.ZodObject<{
        l: z.ZodNumber;
        c: z.ZodNumber;
        h: z.ZodNumber;
        alpha: z.ZodOptional<z.ZodNumber>;
    }, z.core.$strip>>;
    token: z.ZodOptional<z.ZodString>;
    value: z.ZodOptional<z.ZodString>;
    use: z.ZodOptional<z.ZodString>;
    states: z.ZodOptional<z.ZodRecord<z.ZodString, z.ZodString>>;
    intelligence: z.ZodOptional<z.ZodObject<{
        suggestedName: z.ZodString;
        reasoning: z.ZodString;
        emotionalImpact: z.ZodString;
        culturalContext: z.ZodString;
        accessibilityNotes: z.ZodString;
        usageGuidance: z.ZodString;
    }, z.core.$strip>>;
    harmonies: z.ZodOptional<z.ZodObject<{
        complementary: z.ZodObject<{
            l: z.ZodNumber;
            c: z.ZodNumber;
            h: z.ZodNumber;
            alpha: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strip>;
        triadic: z.ZodArray<z.ZodObject<{
            l: z.ZodNumber;
            c: z.ZodNumber;
            h: z.ZodNumber;
            alpha: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strip>>;
        analogous: z.ZodArray<z.ZodObject<{
            l: z.ZodNumber;
            c: z.ZodNumber;
            h: z.ZodNumber;
            alpha: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strip>>;
        tetradic: z.ZodArray<z.ZodObject<{
            l: z.ZodNumber;
            c: z.ZodNumber;
            h: z.ZodNumber;
            alpha: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strip>>;
        monochromatic: z.ZodArray<z.ZodObject<{
            l: z.ZodNumber;
            c: z.ZodNumber;
            h: z.ZodNumber;
            alpha: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strip>>;
    }, z.core.$strip>>;
    accessibility: z.ZodOptional<z.ZodObject<{
        wcagAA: z.ZodOptional<z.ZodObject<{
            normal: z.ZodArray<z.ZodArray<z.ZodNumber>>;
            large: z.ZodArray<z.ZodArray<z.ZodNumber>>;
        }, z.core.$strip>>;
        wcagAAA: z.ZodOptional<z.ZodObject<{
            normal: z.ZodArray<z.ZodArray<z.ZodNumber>>;
            large: z.ZodArray<z.ZodArray<z.ZodNumber>>;
        }, z.core.$strip>>;
        onWhite: z.ZodObject<{
            wcagAA: z.ZodBoolean;
            wcagAAA: z.ZodBoolean;
            contrastRatio: z.ZodNumber;
            aa: z.ZodOptional<z.ZodArray<z.ZodNumber>>;
            aaa: z.ZodOptional<z.ZodArray<z.ZodNumber>>;
        }, z.core.$strip>;
        onBlack: z.ZodObject<{
            wcagAA: z.ZodBoolean;
            wcagAAA: z.ZodBoolean;
            contrastRatio: z.ZodNumber;
            aa: z.ZodOptional<z.ZodArray<z.ZodNumber>>;
            aaa: z.ZodOptional<z.ZodArray<z.ZodNumber>>;
        }, z.core.$strip>;
    }, z.core.$strip>>;
    analysis: z.ZodOptional<z.ZodObject<{
        temperature: z.ZodEnum<{
            warm: "warm";
            cool: "cool";
            neutral: "neutral";
        }>;
        isLight: z.ZodBoolean;
        name: z.ZodString;
    }, z.core.$strip>>;
    atmosphericWeight: z.ZodOptional<z.ZodObject<{
        distanceWeight: z.ZodNumber;
        temperature: z.ZodEnum<{
            warm: "warm";
            cool: "cool";
            neutral: "neutral";
        }>;
        atmosphericRole: z.ZodEnum<{
            background: "background";
            midground: "midground";
            foreground: "foreground";
        }>;
    }, z.core.$strip>>;
    perceptualWeight: z.ZodOptional<z.ZodObject<{
        weight: z.ZodNumber;
        density: z.ZodEnum<{
            light: "light";
            medium: "medium";
            heavy: "heavy";
        }>;
        balancingRecommendation: z.ZodString;
    }, z.core.$strip>>;
    semanticSuggestions: z.ZodOptional<z.ZodObject<{
        danger: z.ZodArray<z.ZodObject<{
            l: z.ZodNumber;
            c: z.ZodNumber;
            h: z.ZodNumber;
            alpha: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strip>>;
        success: z.ZodArray<z.ZodObject<{
            l: z.ZodNumber;
            c: z.ZodNumber;
            h: z.ZodNumber;
            alpha: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strip>>;
        warning: z.ZodArray<z.ZodObject<{
            l: z.ZodNumber;
            c: z.ZodNumber;
            h: z.ZodNumber;
            alpha: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strip>>;
        info: z.ZodArray<z.ZodObject<{
            l: z.ZodNumber;
            c: z.ZodNumber;
            h: z.ZodNumber;
            alpha: z.ZodOptional<z.ZodNumber>;
        }, z.core.$strip>>;
    }, z.core.$strip>>;
    tokenId: z.ZodOptional<z.ZodString>;
}, z.core.$strip>;